// src/enums/accounts.ts
export enum PurposeType {
    Residential = 'Жилое',
    NonResidential = 'Нежилое',
}

export enum AccountStatus {
    Active = 'Aктивный',
    Closed = 'Закрытый',
}

export enum AccountHolderType {
    Individual = 'Физ. лицо',
    LegalEntity = 'Юр. лицо',
}